refactor: Deleted DTOs from data package
